Let WearableController observe the wallpaper's lifecycle
Trigger updating Android Wear devices only when the wallpaper is active.
